+++ dart/collision/ode/OdeCollisionDetector.hpp
@@ -62,7 +62,8 @@ public:
   virtual ~OdeCollisionDetector();
 
   // Documentation inherited
-  std::shared_ptr<CollisionDetector> cloneWithoutCollisionObjects() override;
+  std::shared_ptr<CollisionDetector> cloneWithoutCollisionObjects() const
+  override;
 
   // Documentation inherited
   const std::string& getType() const override;